got rid of the dummy page and fixed redirects
